requirements.txt
pymongo[srv]==3.11.3
Flask==2.0.1
python-decouple==3.4
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.0